Refactor dependency management
This branch refactors the dependency management in this project to
avoid using a private method from pip.  This method
(parse_requirements) has changed its API in future versions of pip
and is not meant to be used by external projects.
This branch uses the preferred method when using both pip and
setup.py: it puts all the dependencies in setup.py and makes pip
use that.